requirements.txt
flask
pandas
faiss-cpu
sentence-transformers
